Agrego modificaciones al template Order para CRUD

+++ Grupo_MVG_TP_FINAL/src/main/java/ar/edu/unju/fi/tpfinal/repository/IOrderRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import ar.edu.unju.fi.tpfinal.model.Order;
 
-public interface IOrderRepository extends CrudRepository<Order, Long>{
+public interface IOrderRepository extends CrudRepository<Order, Integer >{
 
 }
